First functional version of a layered couchdb charm.
I still have many TODOs. The Charm needs to tell connecting services
about admin credentials, and I have a lot of misc TODOs scattered
throughout the file.

But the Charm stands up a server with a functional version of couch. I
was able to connect to it, login, and create a database and
documents. This juju stuff is pretty cool. :-)