Our Charm now ends the admin party in a snazzier event-driven fashion.
The install hook sets an "installed" state, and also sets as
"admin_party" state.

Other handlers in the code note that we are installed, but are still
in admin party, and take the necessary steps to update the config and
set an "config_updated" state. Our start handler then restarts, and
unsets the config_updated flag.